requirements.txt
alabaster==1.0.0
amqp==5.3.1
anyio==4.10.0
asgiref==3.9.1
astroid==3.3.11
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
billiard==3.6.4.0
black==25.1.0
boto3==1.40.21
botocore==1.40.21
celery==5.2.7
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.10.6
cryptography==45.0.6
dill==0.4.0
Django==5.2.5
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-extensions==4.1
django-filter==25.1
django-ratelimit==4.1.0
django-redis==6.0.0
django-security==0.12.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
docutils==0.21.2
drf-spectacular==0.28.0
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==37.6.0
flake8==7.3.0
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kombu==5.5.4
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.17.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
Pygments==2.19.2
pylint==3.3.8
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-dateutil==2.8.0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==4.6.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
s3transfer==0.13.1
sentry-sdk==2.35.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.3
typing_extensions==4.15.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
urllib3==2.5.0
uuid==1.30
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1